BTMKPL - Do not skip sub-merchant account tests for marketplace_v2 branch
In order to allow builds to pass for the backwards incompatible
Marketplace V2 API we must skip sub-merchan account tests on master. On
the marketplace_v2 or btmkpl* branches however these tests should not
be skipped. To allow merging master in to marketplace_v2 we must skip
the tests if the current branch does not support the marketplace v2 API.