BTMKPL cleanup - Remove SubMerchantAccount API references
We started work on the SubMerchantAccount API on the master branch.
The Python client library has been tagged and released with these
changes. As the API changes the Gateway is not backwards compatible
with these prior releases. Development of the SubMerchantAccount
API has moved onto a branch and will be merged into master when it
is stable. This version of the API can be removed from master.