Accept unicode during param key building
Many Python applications and frameworks use unicode strings wherever
possible.  This makes it a bit awkward to do things like update a
subscription, since the following code raises an error:

    braintree.Subscription.update('subscription_id', {
            "add_ons": {
                "add": [
                    {
                        "inherited_from_id": "addon1",
                    }
                ],
                "remove": [u"addon2"]
            },
        })

(Note the unicode add-on name in the "remove" list).

Specifically, it raises:

    AttributeError: 'unicode' object has no attribute 'items'

from resource.py:24.

Be a bit more permissive by allowing instances of str *or* unicode to be
used as a key, allowing the above code to succeed.