Fix http_strategy.http_do regression
DocumentUpload addition modified the `request_body` argument from a
string to a tuple. This was a breaking change for merchants creating
their own strategies. Fixing it by reverting back to `request_body`
being a string for all requests except for when files are passed in
which at that point it becomes a tuple.

Signed-off-by: Steve Coffey <steve.coffey@getbraintree.com>