BTFT-1427 Add three_d_secure_authentication_id to Transaction.sale
What:
Add the concept of a three_d_secure_authentication_id to the Braintree
API, exposing the ID when starting an authentication, and accepting one
later in Transaction.sale.

Why:
The original motivation for this change: currently a nonce cannot be
enriched with both a CVV and 3DS information. For example, if a CVV is
added to a nonce (N1), and then verifyCard is performed on nonce N1,
then a new nonce N2 is returned, which consumes the N1 nonce and thus we
lose the CVV data. With this change, a CVV-only nonce (N3) can be
created, then a separate 3DS-enriched nonce (N4) is created via
verifyCard. Then, the Transaction.sale call can occur with nonce N3
assigned with three_d_secure_authentication_id from N4.