[3893] Add validations for arguments in Address gateway API calls
Why This Change Is Necessary
========================================================================
We received a bug bounty report of a user being able to retrieve
addresses of different customers via Address.delete, Address.find, and
Address.update calls. This is possible by supplying dot notation in the string value of
address IDs.

What These Changes Do To Address the Issue
========================================================================
Adds a "private" method to validate the characters supplied as string
arguments for these method calls.

It's confirmed in the developer documentation that Address IDs can only
be two letters long and alphanumeric, so we do not believe we need to
worry about the API changing to allow other characters.

We also shored up test coverage for all methods :D

Side Effects Caused By This Change
========================================================================

None expected. We structured this logic in a way that wouldn't change
existing tests (or failures)

------------------------------------------------------------------------

**Relevant Links:**

  * [Ticket Link][1]

------------------------------------------------------------------------

[1]: https://engineering.paypalcorp.com/jira/browse/SVM-3893

Signed-off-by: Sara Vasquez <saravasquez@paypal.com>