DTBTSDK-2005: Python path traversal vulernability (#451)
* add two options to fix dot traversal problem

* fix spelling and make options easier to read

* two tests to try deleting a customer thru path traversal

* remove unused functionality

* comment out previous fix

* small refactor

* add with statement for session object

* refactor request creation

* add tests for all delete endpoints

* refactor request prepare

* added mocked http int test

* Revert "comment out previous fix"

This reverts commit 8c30a27cdb1693946cf774d3ee2c1a121865b790.

* Revert "[3893] Add validations for arguments in Address gateway API calls"

This reverts commit 177f2fefde5b22ef9f391e11d2542e3c880611bb.

* update changelog

* add test to check that session closes

* [3893] Add validations for arguments in Address gateway API calls

Why This Change Is Necessary
========================================================================
We received a bug bounty report of a user being able to retrieve
addresses of different customers via Address.delete, Address.find, and
Address.update calls. This is possible by supplying dot notation in the string value of
address IDs.

What These Changes Do To Address the Issue
========================================================================
Adds a "private" method to validate the characters supplied as string
arguments for these method calls.

It's confirmed in the developer documentation that Address IDs can only
be two letters long and alphanumeric, so we do not believe we need to
worry about the API changing to allow other characters.

We also shored up test coverage for all methods :D

Side Effects Caused By This Change
========================================================================

None expected. We structured this logic in a way that wouldn't change
existing tests (or failures)

------------------------------------------------------------------------

**Relevant Links:**

  * [Ticket Link][1]

------------------------------------------------------------------------

[1]: https://engineering.paypalcorp.com/jira/browse/SVM-3893

Signed-off-by: Sara Vasquez <saravasquez@paypal.com>

* re-word changelog entry

* remove unnecessary test from test_address

* move int tests to unit tests

Co-authored-by: Holly Stotelmyer <holly.stotelmyer@getbraintree.com>